requirements.txt
flask
redis
python-dotenv
elasticsearch==8.10.0